Add missing import for hamburger menu and theme switcher to homepage
